Refactor: Crypto 클래스 리팩토링 및 테스트 코드 수정
- Crypto 클래스를 Crypto 인터페이스와 SecureCrypto 구현체로 분리
- 패키지 구조 변경 (security -> security.crypto)
- CryptoTest 클래스명을 SecureCryptoTest로 변경하고, 리팩토링된 클래스에 맞게 수정
