+++ core/security/src/main/java/com/threegap/bitnagil/security/crypto/Crypto.kt
@@ -1,0 +1,7 @@
+package com.threegap.bitnagil.security.crypto
+
+interface Crypto {
+    fun encrypt(bytes: ByteArray): ByteArray
+
+    fun decrypt(bytes: ByteArray): ByteArray
+}
